Address gem loading conflict with the appoxy aws gem.
The aws-sdk gem does not conflict in Ruby module name space
but it does conflict with load paths, namely, the root
aws.rb.  This removes aws.rb from the Amazon gem, allowing
the gems to co-exist for folks wishing to migrate from
appoxy aws gem to the aws-sdk gem in an orderly fashion.