Add support for Message finding out its own origin.
Introduce the concept of a message's origin, and introduce a partial implementation of
an AutoScaling originator (partial, in that there's not coverage for all of the data that
auto-scaling will include in the SNS' body).

The origin-selection currently works simply; there's only one originator at present, so
it's checked in-line in Message's initializer, rather than that iterating through the
available originator modules.