Increased the number of Errors retried by the NetHttpHandler.
Previously, only a short white-list of errors were trapped and
retried as networking errors.  This has been expanded to be
all errors raised from the handler (and Net::HTTP) except for
those on a pass-through list.  This should minimize the number of
un-enuemrated errors that don't get retried (like open ssl socket
errors).