Added support for a single configurable region.
You can now specify what AWS region you want to connect to:

    AWS.config(:region => 'us-west-2')

The individual configuration options for each service endpoints are
now deprecated.  If you want to override the endpoint for a service
you can do so in the constructor:

    AWS::S3.new(:region => 'us-east-1')

    # this also works
    AWS::S3.new(:endpoint => 's3-us-east-1.amazonaws.com')

Documentation changes were made to remove mention of the now
deprecated options.