Added a basic repl executable as aws-sdk-repl.
This repl will default to pry when available, but will
fall back to irb when not available.  It accepts a number
of runtime configuration options and defaults can be exported
to ENV.

By default client requests are logged (in color).  A quick
example:

    aws-sdk-repl
    AWS> s3.buckets['aws-sdk'].objects['key'].content_length
    [AWS S3 200 0.875184 0 retries] head_object(:bucket_name=>"aws-sdk",:key=>"key")
    => 5242881

You can use all methods of the AWS module directly when using Pry.