AWS::DynamoDB::Client now supports both Amazon DynamoDB API versions.
`AWS::DynamoDB::ClientV2` is now deprecated.  Instead of using this class, you can
construct a client of a specific version like so:

    ddb = AWS::DynamoDB::Client.new(:api_version => '2012-08-10')

Constructing a client without an API version returns a client of the oldest
version.  This ensures existing users of AWS::DynamoDB::Client will not experience
runtime errors from the differences.

This works by turning AWS::DynamoDB::Client.new into a factory method.  It returns
an instance of Client::V20111005 or Client::V20120810 based on the :api_version
option.