AWS::Record::HashModel now accepts mutable hash keys.
The default behavior for a class extending `AWS::Record::HashModel` is to
have a managed `#id` attribute.  This attribute is auto-generated before
creation.  It is set to a random UUID.

You can now override this behavior with the `:hash_key` option.  When you do this,
the HashModel's hash key will no longer be managed (no default values).