Added two more Plugin handler helper methods.
`Plugin.request_handler` and `Plugin.response_handler` both
accept a block and are called with the request context or
the response respectively.

Example:

    MyPlugin < Seahorse::Client::Plugin

      request_handler do |context|
        # do something with the request context
      end

      response_handler do |response|
        # do something with the response, this block
        # is called then the response is signaled complete
      end

    end

Without using the new helper methods, the plugin above would look like:

    MyPlugin < Seahorse::Client::Plugin

      handler do |context|
        # do something with the request context
        @handler.call(context)
      end

      response_handler do |context|
        @handler.call(context).on_complete do |response|
          # do something with the response, this block
        end
      end

    end

It is important to note that the .handler method requires you to call the
next handler in the stack.  The `.request_handler` and `.response_handler`
methods take care of this detail on your behalf.