Resolved an issue where config might get stuck in a recursive loop.
When building a service endpoint, the region is used to construct the
default endpoint.  When attempting to determine the region, the endpoint
is is parsed (recursive loop).

The recent patch to make endpoint use the svc region instead of the global
region uncovered this bug.  This patch stops the loops by changing the logic
in #svc_endpoint to used only the explicitly configured regions, and never
a region guessed from the endpoint.