Added full specs for s3 head object using the fixture hander.
In order to make this operation work, there was quite a bit of work
done to the rest protocol plugin to support serializing and parsing
headers, and custom priorities to ensure the uri was populated.

Also added a s3 regional endpoint plugin to handle the special case
of s3 classic endpoint.