DRY'd up the Xml::Handler and Json::Handler classes.
There was nearly 100% overload, so these have been reworked to use
a base Aws::Handler class.  Each subclass now only defines the
`#populate_headers` method and specifies which builder and parser to
use.

The JSON and XML builders and parsers have been updated to have the
same public signatures.  This allows the base handler to not care
what it is working with.