Added OutputShape (extends from InputShape) and a few helpers.
The `InputShape` and `OutputShape` classes now provide a
`#body_member` class instead of `#body_members`.  This method
extracts the need to determine if there is a single payload
member or a hash of them.  In the latter case, they are wrapped
in a StructureShape which inherits metadata and serialization name
from the input/output object.