Now addressing result wrappers in AWS Query services.
Services like IAM, SES, SQS, etc have an extra wrapper
around their response structures.  This was preventing the
XML parser from finding the data in the response body.

A wrapper has been added, as well as a description of
the `ResponseMetadata` -> `RequestId` elements that appear
in the root of the response.