Fixed a bug with AutoScaling scheduled actions.
The `AWS::AutoScaling::ScheduledAction` class incorrectly assumed that their
group names were unique and suffient to identify the action.  In order to correct
this bug, it was required to add a new positional required argument to the
`ScheduledAction` constructor.  This affects two code paths.

First, the scheduled action collection must now be scoped by a group:

    # now raises an error, must scope to a group first
    auto_scaling.scheduled_actions['name']

    # this works
    auto_scaling.groups['group-name'].scheduled_actions['name']

    # this also works
    auto_scaling.scheduled_actions.filter(:group => 'group-name')['name']

Also the constructor now requires a group:

    # no longer works, raises error
    AWS::AutoScaling::ScheduledAction.new('name')

    # this works
    group = AWS::AutoScaling::Group.new('group-name')
    AWS::AutoScaling::ScheduledAction.new(group, 'name')