Added a simple response pager.
The pager is constructed with a response and a paging
configuration.  It will then attempt to fetch the next
page of results based on the paging tokens present
in the response data.

    page = ResponsePager.new(response)
    page.last_page?(false)

    # sends a request to recieve the next response page
    page = page.next_page
    page.last_page?(true)

    page.next_page
    #=> raises ResponsePager::LastPageError

The ResponsePager is a Delegator for the wrapped response.
You can access all of the attributes of the response from
the page object.

    page.data
    #=> { ... }