Initial Support for AWS CloudTrail
AWS CloudTrail is a web service that records AWS API calls for your account and
delivers log files to you. The recorded information includes the identity of the
API caller, the time of the API call, the source IP address of the API caller,
the request parameters, and the response elements returned by the AWS service.