Converted autoload paths from relative to aboslute paths.
When loading the `aws-sdk` gem in the same project as the
`aws-sdk-core` gem, the relative 'aws/errors' and 'aws/version'
paths conflict between the two libraries.

By expanding the relative paths to full paths, autoload no longer
has to search the autoload include paths for the relative autoload
files.  This eliminates the problem of picking up the wrong file
and not finding the autoload constant.

A similar change is being made in `aws-sdk` so that the libraries
can be required in any order.

Fixes #15