Fixed a bug in the S3 signer when signing buckets that have names with dots.
Resolves the bug by correctly using the bucket name to
determine if it is DNS compatible.  Was using the string
literal "bucket" based on the param name.

Added two integration tests to ensure there is no regression.
This tests buckets with dot names using SSL and non-SSL endpoints.
This is because DNS compatible bucket names with/without dots
are handled differently (path vs subdomain).

Fixes #19