Use OpenSSL::Digest where possible.
Per #525 it appears that the regular Digest module is not thread-safe. Testing
showed that OpenSSL::Digest on the other hand is working fine.

Currently the SQS related code still uses Digest in one place. I tried
replacing that but this triggered spec failures for unknown reasons. The
OpenSSL::Digest API should be identical to Digest so I'm not really sure what's
going on here.