No longer serializing `nil` query params to the request URI.
The normalized branch which was merged to mainline changed the
behavior on how `nil` request parameters were treated during
serializations. Previously, these params were omitted if they
were to be serialized to the request querystring. The normalized
branch changed this behavior to always serialize given params,
`nil` or not.

This commit reverts this change of behavior and added a protocol
test to the rest-xml and rest-json handlers to track this behavior.

Fixes #76