Addressed a thread-safety issue in Structure.new
The Structure class is a light-weight wrapper around
the Struct class in Ruby stdlib. The `.new` method of
Structure defines a struct class (as does Struct.new),
but it also registers that class. This ensures multiple
calls to Structure.new with the same member names returns
the same struct class.

This method was generating NoMethodError: undefined method `new' for
nil:NilClass in a multi-threaded environment because of the
shared class map.

This commit adds a mutex around structure class construction
and assignement to the class map.

Fixes #86