Now bypassing the Aws::ClientFactory class.
Removed the concept of runtime resolution of API version
to client class.

In the future, if a service adds a backwards-incompatible
API version, it will be added to a separate service module/client.
This ensures users are not broken at runtime.

This resolves the issue of API updates breaking customers that
did not lock to a service-specific API version. This feature was very
difficult to document and to explain. This was particularly
troublesome in that it defaulted always to the latest API
version, possibly breaking users without their knowledge.

Most users should not be affected by this update unless they
were using API version locking. Simply remove the `:api_version`
constructor param to update.

Updates:

  * Versioned client classes have been removed,
    E.g. `Aws::S3::Client::V20060301` is now simply
    `Aws::S3::Client`.

  * Removed Aws.config API version locking, instead you should
    use the appropriate namespace when multiple versions
    exist.