Removed signature version logic from the region to endpoint logic.
Removed the ability to override the signature version in the
endpoints.json configuration file. Instead the signature
version is fully defined in the service API model metadata.

S3 now defaults to sigv4 for all new regions. For existing
public regions, it will continue to default to the `v3` S3
signature.