Added a plugin that simplifies working with DynamoDB attribute values.
This plugin changes the default format of request params and
response data that represents an Amazon DynamoDB attribute value.
Prior to this plugin, you would have to serialize request
parametes with their type, e.g.

    { s: 'string-value' }
    { n: "5.0" }

This plugin changes `Aws::DynamoDB::Client` operations to accept
and return simplified ruby-name values, such as:

    "string-value"
    5.0

This affects **every** DynamoDB request and response
structure that accepts or returns an attribute value.

This is a backwards-incompatible change. To revert to
the previous behavior, toggle simple attributes off:

    # disable this new default behavior
    Aws::DynamoDB::Client.new(simple_attributes: false)