Separated documentation from the API models.
The primary reason to move the documentation from the API
JSON model files was to reduce the file of the published gem.
With this change, the gem file is reduced from ~740KB down
to ~280KB.

This change also has the benefit that API doc updates will not create
as much churn into the model files. Additionally, it will be easier
to audit what API params have no documentation.