Now requesting S3 url-encode all keys by default.
Amazon S3 supports uploading objects with keys that can not
be safely returned in XML responses, such as ascii 0-10.
These keys cause the SDK to fail when attempting to parse
XML if one such key is present.

To support these keys, S3 accepts a Encoding-Type header. When set,
S3 will url encode all keys in the response XML.

This commit adds a plugin, that sets this request parameter by
default. It then also url-decodes keys automatically before
returning the response data. This makes it so the key handling
is hands-off.

If the user specifies the `:encoding_type` request parameter,
then the plugin is disabled, and the decoding of the keys is
left as an exercise for the user. This is required to be
backwards compatible. This also ensures that if S3 specifies
new encoding types in the future, this will not break.

Fixes #95