Don't Call Dir.home in Ruby 1.8
When attempting to build a SharedCredentialFileProvider in Ruby < 1.9,
you would see a confusing error when Dir.home raises a NoMethodError.

This change checks your RUBY_VERSION and raises a more user-friendly
ArgumentError in that case. It will also prevent a premature crash of
the default credential provider chain when running Ruby 1.8.

Resolves GitHub Issue #604