Now returning a `DefultList` class instead of `nil` for missing XML lists.
This preserves backwards compatability with previous versions, and
eliminates the need for users to nil check prior to accessing list data
that might be missing from the response XML.

To ensure these empty lists are not serialized as request params, they
respond to `#nil?` with `true` when empty. This allows serializers to
safely ignore them, as well as the param converters.