Removed the OperationNames plugin.
This is now core functionality of Seahorse::Client::Base. This means
there is no more runtime patching of the client class. The previous
approach was problematic as once the plugin was applied, it was
impossible to cleanly remove.

Removing this plugin also removed the need for a client class
mutex, as the class is no longer being modified during the
initialization of the first instance.