Fixed resource validator specs.
Since the restructure, the resource definition validator specs
have not been running due to the changed path to fixture files.
Moved the fixtures to the correct location and added a sanity
check that raises an error if the fixture files are missing to
prevent this from happening in the future.

This significantly increases the coverage of the aws-sdk-resources
gem.