Patched buggy S3 waiter definitions.
A suitable fix needs to be merged to the upstream definition file.
Currently, the definition assumes a particular error code when S3
does not return one. S3 does not return an error code as these
are HEAD requests with XML bodies in the response.

Ideally the waiter configuration would support failure/success acceptors
that match on the status code (404 in this case).

Until then, this changes the error codes to those generated by this SDK.