Resolved an issue with sigv4 signatures.
The normalized query params were being sorted by "name=value".
This caused an issue when two params were named as such:

    "q=value"
    "q.options=value"

The sort cause q.options to be first, when it should have been
q, because of the precedence of =. This fix ensures they are
sorted by name only, and that two params with the same name
do not change places.

Fixes #112