Resolves a validation error with IAM.resources.json
The User#create_access_key method was attempting to return
an object with data, but the data path did not resolve to the
appropriate shape. It was attempting to return an "AccessKey" shape
instead of the "AccessKeyMetadata" shape required by its type.

To resolve this, a new resource type has been added, "AccessKeyPair".
This resource has the access key id AND the secret access key.

    # provides access to the secret only on newly created objects
    key_pair = IAM::Resource.new.user('name').create_access_key
    key_pair.access_key_id
    key_pair.secret_access_key

    # secrets are not retrievable after creation
    key = IAM::Resource.new.user('name').access_key('id)
    key.access_key_id
    key.secret_access_key #=> raises NoMethodError