All operations now accept a `:response_target` request param.
This optional parameter is now used to set the target of
the HTTP response body. Normal defaults still apply, but this
makes it simpler for streaming download operations, such
as S3::Client#get_object and Glacier::Client#get_job_output
to specify a file to download the response body to:

    # download an object to a file
    s3.get_object(bucket:'aws-sdk', key:'key', response_target:'path/to/file')

    # this still works
    s3.get_object({bucket:'aws-sdk', key:'key'}, {target:'path/to/file'})

Also, the new `:response_target` target parameter is now documented
on all API calls, and is included in the formatting example for
streaming operations (those named above).

Fixes #55