Improved instance credential refresh logic.
When instance profile credentials are near expiration, it was possible
previously for multiple threads to trigger a refresh. While to code to
update the credential values was behind a mutex, it was possible for
multiple calls to be made.

Resolved this by verifying the credentials were not updated while
waiting to aquire the lock.

    def refresh_if_near_expiration
      if near_expiration?
        @mutex.synchronize do
          _refresh if near_expiration?
        end
      end
    end