Resolved an issue with Aws::CloudFront::Client paging.
The client was requesting pager using the operation name
that had a YYYY_MM_DD suffix. This caused the client paginator
to return a null paginator.

Now removing this suffix from the Cloud Front operaiton names and
added tests to ensure the paginator for "ListDistributions" is not
a null pager.

Fixes #169