Empty set. How to be?
Hi everyone!

Is it possible to cast empty set as empty list? Or shall I handle empty sets in my client code? 
I propose to cast the empty sets as empty lists because this is a good way to follow simple rule: "What You See Is What You Get"
We see the empty set so this is the same as the empty list, isn't it?

Referring to php-sdk, php has no set class but only array and all operations with sets performed over arrays. So under php context empty sets and empty lists are meaning same things.

Thanks,
Anton