Don't Add Handlers When Operation Filter is Empty
Within Seahorse, when an operation filter was an empty array, the
behavior defaulted to adding that handler to all operations. This is not
the correct behavior, for example this meant we added the Pagination
handler to ALL operations for services that had NO pagable operations
defined. This change allows a `nil` operation list as the way to specify
adding a handler to all operations, while an empty operation list will
mean the handler is added to no operations.

Resolve GitHub issue #984.