Add a Blacklist of Some Headers From Signing
Certain headers beyond the authorization header could, in theory, be
changed for various reasons (including but not limited to proxies) while
in transit and after signing. This would lead to signature errors being
returned, despite no actual problems with signing.

This change is a first draft loosely based off the header blacklist used
in the AWS SDK for PHP.