Resolved an issue with EC2 request ids.
This is a regression from when we switched to stream parsing the XML.
Older versions of the SDK parsed the entire XML document into a DOM,
and then walked that DOM to parse the values. Because the entire
document was no longer available in memory, the parser could no
longer yield it as a hash. This caused the EC2 protocol handler
to silently fail, no longer able to populate the request ID.

This update adds a callback to the XML parser that yields when
un-modeled/un-handled XML values are present in the response. The
EC2 handler registers a callback and checks the context/path
of the unhandled value and captures it if it is the request id.

To prevent a regression in the future, a unit test was added to ensure
the EC2 client correctly extracts the request id from the HTTP response
body and stores it in the response context.

Fixes #1038