Get Correct sigv4_region When No endpointPrefix
Aws::IoTDataPlane::Client intentionally does not define `endpointPrefix`
in the configured API metadata as a mechanic to force the user to supply
an `:endpoint` value on client construction.

The previous attempted fix failed to note that, in absence of
`endpontPrefix`, all endpoints will match with the first selector,
hard-coded to "us-east-1". This change will check for this case, and use
your client configured region when it occurs.

* Resolves GitHub Issue #1070
* Related to GitHub Issue #1052
* Fixes Regression In GitHub Pull Request #1053