Avoid Double URL Encoding S3 Keys
When performing a copy operation with the S3 Object resource, the object
key would be URL encoded before being sent on to the S3 client, which
URL encodes the key a second time. This was causing a 404 to
short-circuit the copy operation.

Tests were checking the inputs into the client call, which would not
catch the second step of the double encoding.