Fix for Aws::S3::Client#head_bucket with improper region.
Resolved the following issues:

  s3 = Aws::S3::Client.new(region: 'us-west-1')
  s3.head_bucket(bucket: 'us-west-2-bucket')
  #=> raises Aws::S3::Errors::BadRequest (without a message)

If the bucket exists, but you are configured for the incorrect region,
the SDK fails to extract the actual bucket region from the 400 response.
This results in an empty error message. Added support for loading the
bucket region from the "x-amz-bucket-region" header, which is necessary
for all HEAD responses.

Fixes #1161