Fix for Aws::Xml::Parser concurrency issue.
The XML parser supports multiple backends. There is a race condition
when loading the parser code where the default parser has not been
chosen, but the first instance of the parser has been constructed. In
this scenario, it raises with a undefined method error on `nil`. This is
because the self.class.engine is returning `nil`.

Here is a minimal test case that reproduces this issue. It reproduces
reliably under Ruby 1.9.3:

```ruby
require 'aws-sdk-core'
require 'thread'

sqs = Aws::SQS::Client.new
threads = 10.times.map do
  Thread.new do
    sqs.send_message(queue_url:"https://...", message_body:"string")
  end
end
threads.map(&:join)
```

The work-around forces the constructor to set the default engine if
this hasn't happened yet.

Fixes #1135