Code Review Feedback and Added Credential Support
Brings this change up to spec for credentials, assume role, and region
support. Includes code review feedback to this point, and relevant test
improvements.

Still to be done is the addition of the Amazon Simple Storage Service
specific configuration values, and the wiring of Assume Role to the
credential provider chain.