Fix class organization headings in docs
The SDK is doing some organization to the API docs page (http://docs.aws.amazon.com/sdkforruby/api/) class list. Clicking on the headings results in being redirected to an "undefined" page because YARD is adding links to the lis programatically. This PR adds a small JS file to remove the linking from the heading lis.