Introduced the concept of generator client plugins.
To support documenting plugin options during code generation,
plugins are now passed to the generator. The generator opens the
plugin ruby file and queries the plugin for its documentation.

Plugins are now managed as part of the service manifest file. You
can add and remove plugins that are merged with the default plugin
list.