Large credential and signer refactor.
Extracted a sigv2 signing gem to support simple db. A client will now
also re-use a single singer. This now makes it possible for the sigv4
signer to case the derived key.

Also separated the credential configuration logic from the sigv4 plugin
so it can be used by non-sigv4 services.