Fix for the AutoScaling resource model.
There was a bug in the #load definition for AutoScaling tags.
It was calling DescribeTags and filtering only by the tag key and value,
but not by the resource type and id. This allowed it to receive
the wrong tags back in the response.

Additionally, there was a related issue where extra calls were being
made to DescribeTags when the data was present already on the
AutoScaling group. Added the missing path key on the resource model,
and now there are no additional describe calls being made.

Fixes #1313