Fix Mutation Bug in DynamoDB Simple Attributes
Under a particular set of circumstances, a return value struct can have
SimpleAttributes transformations applied multiple times, which can case
incorrect types and values to be written (or more likely, fail to be
written).

This doesn't happen for hashes, but we hadn't applied the same approach
to structs. This change applies the same approach to type structs,
including a previously failing test case.

Resolves GitHub issue #1406