Set Thread#abort_on_exception on spawned threads
By default threads won't propagate exceptions that occur inside them to
the main thread, unless we call Thread#value on a thread that exited
with an exception, in which case its exception will be raised in the
main thread. Because some folks might not know that last bit, for
readability we explicitly set Thread#abort_on_exception on spawned
threads.