Use Digest#base64digest method
Other than being shorter, it has two qualities over Base64.encode64 that
are desirable in this case. One is that it doesn't include a trailing
newline at the end of the result. Other one is that it doesn't wrap the
result in 60 columns, meaning it doesn't add a newline after each 60
characters.

A general alternative to Base64.encode64 is Base64.strict_encode64,
which should have the same desirable behaviour, but calling
Digest#base64digest is shorter.