Reduce memory allocation in checksums and signatures
When IO is #read without a buffer string, it allocates a new string for
each read chunk. Even though by reading the file in chunks we keep
memory usage low, we still give GC work to clean up all those strings.

However, if we pass an additional buffer string to #read, instead of
allocating new string for each chunk, each chunk will be read into the
existing buffer string (replacing any previous value). That means that
the buffer string is the only string that's allocated.

The IO body already has to accept the additional buffer string, because
Net::HTTP uses IO.copy_stream to stream the IO to the TCP socket, which
will throw an exception if the IO body only accepts the length argument.