Handle unlinked files in formatter
On many operating systems (e.g. Linux and OSX), an unlinked file is perfectly usable as long as it is open, but it does no longer have a path, and `File.size(path)` would fail as it would receive a `nil` path

The test case creates a temporary file and opens a normal file reference to it, after which the temporary file is unlinked. In modern versions of Ruby, we could use `Tempfile.create` for this purpose, but this isn't available in older rubies.