require `cluster` for ECS DescribeTasks
`Aws::ECS::Client#describe_tasks` will permit an API request to be sent
without the `cluster` request parameter, but such a request will always
be met with a response containing a populated `failures` array. The
reason given for each failure is `MISSING`.

```
> ecs_client.describe_tasks(tasks: %w[a8d82dbd-4796-451b-bc8d-21ef77285d9f])
  failures = [
    [0] #<#<Class:0x00007f85114428a0>:Aws::ECS::Types::Failure:0x7f8511f3d408
      arn = "arn:aws:ecs:us-east-1:963710170562:task/a8d82dbd-4796-451b-bc8d-21ef77285d9f",
      reason = "MISSING"
    >
  ],
  tasks = []
```

When the `cluster` request parameter is included and set equal to a
valid value (the arn for a cluster that the submitted task arns are
associated with), the response comes back as desired without any
failures.

It seems that the documentation should be updated to reflect that
`cluster` is in fact a request request parameter.