Create .rspec
Warnings are an important part of Ruby's upgrade path. They indicate potential bugs or other issues with the code (for example, typos in `@instance_variable` names, clobbering functions and so on.) I suggest that this is enabled and that the code is fixed so that it no longer produces warnings.