Remove Final Instances of Error Output
It does seem more reasonable to have AsyncClient instances emit this
error, but since they are eager loaded, this in effect has the error
appear anytime someone loads `aws-sdk-kinesis` or
`aws-sdk-transscribestreamingservice`, and the former case especially is
a bit of a regression.

Since AsyncClient creation will always generate an exception if the gem
isn't present, I'm removing these also.

Resolves #1998