Update MINIMUM_CORE_VERSION to 3.52.0 (#2047)
In recent update a new plugin transfer_encoding was added to sdk-core.
This plugin is in use in many gems which are using sdk-core.
But minimum version of sdk core was stayed on old version 3.48.2.
Since we are introduced a new pluging and used it everywhere we have to
bump minimal version to 3.52.0. Because new plugin was introduced in
this version.