Fix erroneously reaped sessions from `Seahorse::Client::NetHttp::ConnectionPool` due to bad `last_used` time calculation
To detect which sessions are stale and remove them from the pool the `_clean` method is comparing the current and `session.last_used` timestamps. If the difference is larger than the predefined `http_idle_timeout` the session is considered stale and is terminated.

The issue here is that the `session.last_used` value is set to be the time right before a request is made thus wrongfully including the request duration in the idle time equation consequently terminating the session instead of reusing it. This is often manifested when a slow connection causes a request (e.g. uploaded part) to take longer than the configured `http_idle_timeout`.

Moving the time sampling after the request is complete fixes the issue and is in line with how Ruby `Net::HTTP` is doing its internal `keep_alive_timeout` calculations [1].

[1] https://github.com/ruby/ruby/blob/af11efd377965b6601bb54aa79072ef0789dc525/lib/net/http.rb#L1598